Change Directory Structure of Model
