+++ lib/model/weather/weather_request.dart
